+++ src/main/java/com/wildcodeschool/giftmefive/controller/ControllerLists.java
@@ -17,7 +17,7 @@ public class ControllerLists {
         return "lists";
     }
     @GetMapping("/listes/delete")
-    public String deleteList(@RequestParam int id) {
+    public String deleteList(@RequestParam long id) {
         listsRepository.deleteGiftFromList(id);
         listsRepository.deleteList(id);
         return "redirect:/listes";
