+++ src/main/java/com/wildcodeschool/giftmefive/repository/ListsRepository.java
@@ -24,12 +24,12 @@ public class ListsRepository {
             ResultSet resultSet = statement.executeQuery();
             List<ListGift> lists = new ArrayList<ListGift>();
             while (resultSet.next()) {
-                int idList = resultSet.getInt("id_list");
+                long idList = resultSet.getLong("id_list");
                 String listName = resultSet.getString("list_name");
                 String description = resultSet.getString("description");
                 String urlImage = resultSet.getString("url_image");
                 String urlShare = resultSet.getString("url_share");
-                int idUser = resultSet.getInt("id_user");
+                long idUser = resultSet.getLong("id_user");
                 lists.add(new ListGift(idList, listName, description, urlImage, urlShare, idUser));
             }
             return lists;
@@ -38,7 +38,7 @@ public class ListsRepository {
         }
         return null;
     }
-    public ListGift findById(int idList) {
+    public ListGift findById(long idList) {
         try {
             Connection connection = DriverManager.getConnection(
                     DB_URL, DB_USER, DB_PASSWORD
@@ -54,14 +54,14 @@ public class ListsRepository {
             String description = resultSet.getString("description");
             String urlImage = resultSet.getString("url_image");
             String urlShare = resultSet.getString("url_share");
-            int idUser = resultSet.getInt("id_user");
+            long idUser = resultSet.getLong("id_user");
             return new ListGift(idList, listName, description, urlImage, urlShare, idUser);
         } catch (SQLException e) {
             e.printStackTrace();
         }
         return null;
     }
-        public void deleteList(int id) {
+        public void deleteList(long id) {
             try {
                 Connection connection = DriverManager.getConnection(
                         DB_URL, DB_USER, DB_PASSWORD
@@ -69,7 +69,7 @@ public class ListsRepository {
                 PreparedStatement statement = connection.prepareStatement(
                         "DELETE FROM list WHERE id_list=?;"
             );
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             if (statement.executeUpdate() != 1) {
                 throw new SQLException("failed to delete data");
             }
@@ -77,7 +77,7 @@ public class ListsRepository {
             e.printStackTrace();
         }
     }
-    public void deleteGiftFromList(int id) {
+    public void deleteGiftFromList(long id) {
         try {
             Connection connection = DriverManager.getConnection(
                     DB_URL, DB_USER, DB_PASSWORD
@@ -85,7 +85,7 @@ public class ListsRepository {
             PreparedStatement statement = connection.prepareStatement(
                     "DELETE FROM gift WHERE id_list=?;"
             );
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             if (statement.executeUpdate() != 1) {
                 throw new SQLException("failed to delete data");
             }
